pipe semi fix

+++ Kernel/interruptions/syscallDispatcher.c
@@ -73,7 +73,7 @@ static int memDumpWrapper(uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t r8,
 }
 
 static int newProcessWrapper(uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t r8, uint64_t r9) {
-    return newProcess((void (*)(int, char **)) rsi, (int) rdx, (char **) rcx, (uint8_t) r8, (uint16_t *) r9);
+    return newProcess((void (*)(int, char **)) rsi, (int) rdx, (char **) rcx, (int) r8, (int *) r9);
 }
 
 static int killProcessWrapper(uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t r8, uint64_t r9) {
